requirements.txt
torch
torchvision
timm
datasets
huggingface_hub
pyyaml
tqdm
pandas
numpy
pillow
scikit-learn
matplotlib
requests
pytest
timm